requirements.txt
streamlit==1.33.0
pandas==2.2.1
python-dotenv==1.0.1
beautifulsoup4==4.12.3

# LangChain ecosystem (PINNED to working versions)
langchain==1.2.3
langchain-core==1.2.7
langchain-community==0.4.1
langchain-openai==1.1.7
langchain-chroma==1.1.0
chromadb==1.4.0
langchain-text-splitters

# OpenAI + tokenization
openai==2.15.0
#tiktoken==0.5.2

# Embeddings (local)
sentence-transformers==2.6.1

# Fix httpx proxies issue (keep pinned)
httpx==0.27.2
